# openapi.yaml
openapi: 3.0.0
info:
  title: HonestNeed API
  version: 1.0.0
  description: |
    HonestNeed Platform API - Complete specification for campaign management, user authentication, and transactions.
    
    ## Authentication
    All protected endpoints require JWT token in Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Error Handling
    All API responses follow standard format:
    ```json
    {
      "success": true|false,
      "data": {} || null,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable message",
        "details": {}
      }
    }
    ```
    
    ## Development
    - Base URL (dev): http://localhost:5000
    - Base URL (staging): https://staging-api.honestneed.com
    - Base URL (prod): https://api.honestneed.com
  contact:
    name: HonestNeed Support
    email: support@honestneed.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://staging-api.honestneed.com
    description: Staging server
  - url: https://api.honestneed.com
    description: Production server
paths:
...
  /api/v1/transactions/{transactionId}:
    get:
      summary: Get transaction details
      tags:
        - Transactions
      operationId: getTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '401':
          description: Unauthorized
        '404':
          description: Transaction not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        campaign_id:
          type: string
        user_id:
          type: string
        amount:
          type: integer
          description: Amount in cents
        transaction_type:
          type: string
          enum: [donation, share_reward, refund, platform_fee]
        payment_method:
          type: string
          enum: [stripe, paypal, bank_transfer]
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        payment_id:
          type: string
          description: External payment processor ID
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TransactionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Transaction'
...